Added additional properties in the models

+++ src/Via.Movies.Api/Models/Movie.cs
@@ -2,7 +2,7 @@
 
 public class Movie
 {
-    public required int Id { get; set; }
-    public required string Title { get; set; }
-    public required long Year { get; set; }
+	public required long Id { get; set; }
+	public required string Title { get; set; }
+	public required int Year { get; set; }
 }
